engine/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
boto3==1.40.10
botocore==1.40.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
griffe==1.11.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
mcp==1.12.4
moviepy==1.0.3
numpy==2.3.2
openai==1.99.9
openai-agents==0.2.6
pillow==11.3.0
proglog==0.1.12
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
tqdm==4.67.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
